+++ lib/src/words_api.dart
@@ -135,7 +135,7 @@ class WordsApi {
     return await _apiClient.call(request) as WordsResponse;
   }
 
-  /// Supported extensions: ".doc", ".docx", ".docm", ".dot", ".dotm", ".dotx", ".flatopc", ".fopc", ".flatopc_macro", ".fopc_macro", ".flatopc_template", ".fopc_template", ".flatopc_template_macro", ".fopc_template_macro", ".wordml", ".wml", ".rtf".
+  /// Supported all save format extensions.
   Future< DocumentResponse > createDocument(CreateDocumentRequest request) async {
     return await _apiClient.call(request) as DocumentResponse;
   }